CD command working with strategies pattern
